Separate the find ref test and separate the tests by data provider
